Update server, config, models, routes and node-modules to incorporate passport
