--- passport/package.json
@@ -1,28 +1,0 @@
-{
-  "name": "passport-tutorial",
-  "version": "1.0.0",
-  "description": "",
-  "main": "app.js",
-  "dependencies": {
-    "body-parser": "^1.18.3",
-    "cors": "^2.8.4",
-    "errorhandler": "^1.5.0",
-    "express": "^4.16.3",
-    "express-session": "^1.15.6",
-    "express-jwt": "^5.3.1",
-    "jsonwebtoken": "^8.2.2",
-    "mongoose": "^5.1.3",
-    "morgan": "^1.9.0",
-    "passport-local": "^1.0.0",
-    "passport": "^0.4.0",
-    "path": "^0.12.7"
-  },
-  "devDependencies": {},
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "nodemon app.js"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC"
-}
