--- passport/routes/api/index.js
@@ -1,6 +1,0 @@
-const express = require('express');
-const router = express.Router();
-
-router.use('/users', require('./users'));
-
-module.exports = router;
