+++ routes/index.js
@@ -1,13 +1,20 @@
-const path = require("path");
-const router = require("express").Router();
-const apiRoutes = require("./api");
+// const path = require("path");
+// const router = require("express").Router();
+// const apiRoutes = require("./api");
 
-// API Routes
-router.use("/api", apiRoutes);
+// // API Routes
+// router.use("/api", apiRoutes);
 
-// If no API routes are hit, send the React app
-router.use(function (req, res) {
-    res.sendFile(path.join(__dirname, "../client/public/index.html"));
-});
+// // If no API routes are hit, send the React app
+// router.use(function (req, res) {
+//     res.sendFile(path.join(__dirname, "../client/public/index.html"));
+// });
+
+// module.exports = router;
+
+const express = require('express');
+const router = express.Router();
+
+router.use('/api', require('./api'));
 
 module.exports = router;
