+++ server.js
@@ -1,24 +1,89 @@
-const express = require("express");
+// const express = require("express");
 
-const mongoose = require("mongoose");
-const routes = require("./routes");
+// const mongoose = require("mongoose");
+// const routes = require("./routes");
+// const app = express();
+// const PORT = process.env.PORT || 3001;
+
+// // Define middleware here
+// app.use(express.urlencoded({ extended: true }));
+// app.use(express.json());
+// // Serve up static assets (usually on heroku)
+// if (process.env.NODE_ENV === "production") {
+//     app.use(express.static("client/build"));
+// }
+// // Add routes, both API and view
+// app.use(routes);
+
+// // Connect to the Mongo DB
+// mongoose.connect(process.env.MONGODB_URI || "mongodb://localhost/shreddit");
+
+// // Start the API server
+// app.listen(PORT, function () {
+//     console.log(`🌎  ==> API Server now listening on PORT ${PORT}!`);
+// });
+
+const express = require('express');
+const path = require('path');
+const bodyParser = require('body-parser');
+const session = require('express-session');
+const cors = require('cors');
+const mongoose = require('mongoose');
+const errorHandler = require('errorhandler');
+
+//Configure mongoose's promise to global promise
+mongoose.promise = global.Promise;
+
+//Configure isProduction variable
+const isProduction = process.env.NODE_ENV === 'production';
+
+//Initiate our app
 const app = express();
-const PORT = process.env.PORT || 3001;
 
-// Define middleware here
-app.use(express.urlencoded({ extended: true }));
-app.use(express.json());
-// Serve up static assets (usually on heroku)
-if (process.env.NODE_ENV === "production") {
-    app.use(express.static("client/build"));
+//Configure our app
+app.use(cors());
+app.use(require('morgan')('dev'));
+app.use(bodyParser.urlencoded({ extended: false }));
+app.use(bodyParser.json());
+app.use(express.static(path.join(__dirname, 'public')));
+app.use(session({ secret: 'passport-tutorial', cookie: { maxAge: 60000 }, resave: false, saveUninitialized: false }));
+
+if(!isProduction) {
+  app.use(errorHandler());
 }
-// Add routes, both API and view
-app.use(routes);
 
-// Connect to the Mongo DB
-mongoose.connect(process.env.MONGODB_URI || "mongodb://localhost/shreddit");
+//Configure Mongoose
+mongoose.connect('mongodb://localhost/shreddit');
+mongoose.set('debug', true);
 
-// Start the API server
-app.listen(PORT, function () {
-    console.log(`🌎  ==> API Server now listening on PORT ${PORT}!`);
+//Models & routes
+require('./models/Users');
+require('./config/passport');
+app.use(require('./routes'));
+
+//Error handlers & middlewares
+if(!isProduction) {
+  app.use((err, req, res) => {
+    res.status(err.status || 500);
+
+    res.json({
+      errors: {
+        message: err.message,
+        error: err,
+      },
+    });
+  });
+}
+
+app.use((err, req, res) => {
+  res.status(err.status || 500);
+
+  res.json({
+    errors: {
+      message: err.message,
+      error: {},
+    },
+  });
 });
+
+app.listen(8000, () => console.log('Server running on http://localhost:8000/'));
